+++ lib/nightcrawler_swift/commands/upload.rb
@@ -6,7 +6,7 @@ module NightcrawlerSwift
       headers = {etag: etag(content), content_type: content_type(file)}
 
       max_age = opts[:max_age] || options.max_age
-      headers.merge!(cache_control: "public, max-age=#{max_age}") if max_age
+      headers.merge!(cache_control: "public, max-age=#{max_age}", expires: expires(max_age)) if max_age
 
       response = put "#{connection.upload_url}/#{path}", body: content, headers: headers
       [200, 201].include?(response.code)
@@ -19,6 +19,10 @@ module NightcrawlerSwift
 
     def etag content
       Digest::MD5.hexdigest(content)
+    end
+
+    def expires max_age
+      CGI.rfc1123_date(Time.now + max_age)
     end
 
   end